rename project helper and move code in client wrapper
